refactor(hooks): extract unknown-repo push gate out of hook_router god-module
hook_router.py is already over the module-health 500-LOC / 10-function
cap, and the ratchet is HEAD-pegged and shrink-only: an over-cap file may
only shrink relative to its prior content, with no bypass. The 144-line
unknown-repo SCOPE push gate this PR added pushed it from 6388 to 6484 LOC
(56 to 57 public functions), tripping the gate.

Move the gate into a new sibling module hooks/scripts/unknown_repo_push_gate.py
and import its single public handler at the top of hook_router. This is a
pure mechanical move with no behaviour change: the regexes, the four
helpers, the deny-message constant, and handle_block_unknown_repo_push are
byte-identical apart from the lazy back-imports of the shared hook_router
helpers (_bootstrap_teatree_django, _resolve_cwd_repo, _fail_open_or_deny,
_teatree_bool_setting), which stay in hook_router (9+ other callers) and are
imported at call time to break the import cycle. The load-bearing ordering --
_bootstrap_teatree_django() before classify_active_push -- is preserved.

hook_router.py drops to 6393 LOC / 56 public functions (shrunk vs HEAD on
both axes, so the ratchet passes); the new module is 126 LOC / 1 public
function, well under the cap. hooks/scripts is added to the ty and pytest
search paths so the static checker resolves the runtime sibling import. The
registry entry stays wired and every gate test stays green.
